tercera práctica

+++ src/dao/mongo/products.mongo.js
@@ -3,13 +3,13 @@ import ProductDTO from '../../dto/product.dto.js';
 import { faker } from '@faker-js/faker/locale/es';
 
 class ProductsMongoDAO {
-	constructor() {}
+	constructor() { }
 
 	async getProductsDao() {
 		try {
 			const products = await productModel.find();
-			if(!products) return `No products found.`
-			return products
+			if (!products) return `No products found.`;
+			return products;
 		} catch (error) {
 			return `${error}`;
 		}
@@ -18,8 +18,8 @@ class ProductsMongoDAO {
 	async getProductDao(pid) {
 		try {
 			const product = await productModel.findById(pid);
-			if(!product) return `No product found with ID '${pid}'.`
-			return product
+			if (!product) return `No product found with ID '${pid}'.`;
+			return product;
 		} catch (error) {
 			return `${error}`;
 		}
@@ -28,8 +28,8 @@ class ProductsMongoDAO {
 	async createProductDao(product) {
 		try {
 			const newProduct = await productModel.create(product);
-			if(!newProduct) return `No product was created.`
-			return newProduct
+			if (!newProduct) return `No product was created.`;
+			return newProduct;
 		} catch (error) {
 			return `${error}`;
 		}
@@ -38,7 +38,7 @@ class ProductsMongoDAO {
 	async updateProductDao(pid, product) {
 		try {
 			const productToModify = await productModel.findById(pid);
-			if(!productToModify) return `No product found with ID '${pid}'.`
+			if (!productToModify) return `No product found with ID '${pid}'.`;
 			await productModel.updateOne({ _id: pid }, product);
 			const updatedProduct = await productModel.findById(pid);
 			return updatedProduct;
@@ -47,15 +47,21 @@ class ProductsMongoDAO {
 		}
 	}
 
-	async deleteProductDao(pid) {
+	async deleteProductDao(req, res, pid) {
 		try {
+			const { user } = req.session;
 			const product = await productModel.findById(pid);
-			if(!product) return `No product found with ID '${pid}'.`
+			if (!product) return `No product found with ID '${pid}'.`;
+			console.log("user.role", user.role)
+			console.log("user.email", user.email)
+			console.log("product.owner", product.owner)
+
+			if (user.role == 'premium' && user.email != product.owner) return `Can't delete product from other owner.`;
 
 			await productModel.deleteOne({ _id: pid });
 			const productDeleted = await productModel.findById(pid);
 
-			if (productDeleted) return `No product was deleted.`
+			if (productDeleted) return `No product was deleted.`;
 			const products = await productModel.find();
 			return products;
 		} catch (error) {
@@ -63,23 +69,25 @@ class ProductsMongoDAO {
 		}
 	}
 
-	async generateProductsDao() {
+	async generateProductsDao(req, res) {
 		try {
-			for ( let i = 0 ; i < 100 ; i++ ) {
+			const { user } = req.session;
+			for (let i = 0; i < 100; i++) {
 				const product = {
 					title: faker.commerce.productName(),
 					description: faker.commerce.productDescription(),
 					code: faker.string.uuid(),
 					price: faker.commerce.price(),
-					stock: faker.number.int({min: 0, max: 100}),
+					stock: faker.number.int({ min: 0, max: 100 }),
 					category: faker.commerce.product(),
-				}
-				const newProduct = new ProductDTO(product)
-				await productModel.create(newProduct)
+					owner: user.email,
+				};
+				const newProduct = new ProductDTO(product);
+				await productModel.create(newProduct);
 			}
 
 			const products = await productModel.find();
-			if(!products) return `No products were created.`;
+			if (!products) return `No products were created.`;
 			return products;
 		} catch (error) {
 			return `${error}`;
